+++ lib/Clarity.ml
@@ -1,6 +1,7 @@
 (* This is free and unencumbered software released into the public domain. *)
 
 #include "Clarity/grammar.ml"
+#include "Clarity/library.ml"
 #include "Clarity/compile.ml"
 #include "Clarity/parse.ml"
 #include "Clarity/print.ml"