+++ lib/Clarity.mli
@@ -1,6 +1,7 @@
 (* This is free and unencumbered software released into the public domain. *)
 
 #include "Clarity/grammar.ml"
+#include "Clarity/library.mli"
 #include "Clarity/compile.mli"
 #include "Clarity/parse.mli"
 #include "Clarity/print.mli"