+++ lib/Clarity/library.mli
@@ -1,0 +1,3 @@
+(* This is free and unencumbered software released into the public domain. *)
+
+val is_primitive : string -> bool